refactor code splitting
