+++ .env_example
@@ -1,2 +1,3 @@
 PORT=1234
 DB_HOST=
+SECRET=
