+++ api/users.js
@@ -1,9 +1,8 @@
-/* eslint-disable indent */
 const { Router } = require('express');
 const router = Router();
 
 const { validate, authenticate } = require('../middleware/');
-const { registrationValidator, updateUser } = require('../utils/validate/schemas');
+const { registrationValidator, updateSubscriptionValidator } = require('../utils/validate/schemas');
 
 const { users: ctrl } = require('../controllers/');
 
@@ -11,6 +10,6 @@ router.post('/signup', validate(registrationValidator), ctrl.signup);
 router.post('/login', validate(registrationValidator), ctrl.login);
 router.get('/logout', authenticate, ctrl.logout);
 router.get('/current', authenticate, ctrl.getCurrentUser);
-router.patch('/', validate(updateUser), authenticate, ctrl.updateSubscription);
+router.patch('/', validate(updateSubscriptionValidator), authenticate, ctrl.updateSubscription);
 
 module.exports = router;
