+++ controllers/contacts/index.js
@@ -1,3 +1,15 @@
-const contacts = require('./contacts.controllers');
+const listContacts = require('./listContacts');
+const getContactById = require('./getContactById');
+const addContact = require('./addContact');
+const updateContactById = require('./updateContactById');
+const deleteContactById = require('./deleteContactById');
+const toggleFavoriteContact = require('./toggleFavoriteContact');
 
-module.exports = { contacts };
+module.exports = {
+  listContacts,
+  getContactById,
+  addContact,
+  updateContactById,
+  deleteContactById,
+  toggleFavoriteContact,
+};
