+++ controllers/users/login.js
@@ -1,14 +1,15 @@
 const jwt = require('jsonwebtoken');
 const { user: service } = require('../../services');
+const HTTP_STATUS = require('../../utils/httpStatusCodes');
 
 const login = async (req, res, next) => {
   const { email, password } = req.body;
   try {
     const user = await service.getOne({ email });
     if (!user || !user.validatePassword(password)) {
-      res.status(401).json({
+      res.status(HTTP_STATUS.UNAUTHORIZED).json({
         status: 'Error',
-        code: 401,
+        code: HTTP_STATUS.UNAUTHORIZED,
         message: 'Email or password is wrong',
       });
     }
@@ -21,9 +22,9 @@ const login = async (req, res, next) => {
 
     await service.updateById(payload.id, { token });
 
-    res.status(200).json({
+    res.status(HTTP_STATUS.SUCCESS).json({
       status: 'Success',
-      code: 200,
+      code: HTTP_STATUS.SUCCESS,
       data: {
         token,
         user: { email: user.email, subscription: user.subscription },
