+++ middleware/validate.js
@@ -1,11 +1,13 @@
+const HTTP_STATUS = require('../utils/httpStatusCodes');
+
 const validate = validator => {
   return async (req, res, next) => {
     const error = await validator(req.body);
 
     if (error) {
-      res.status(400).json({
+      res.status(HTTP_STATUS.BAD_REQUEST).json({
         status: 'Error',
-        code: 400,
+        code: HTTP_STATUS.BAD_REQUEST,
         message: error.message,
       });
       return;
