+++ utils/httpStatusCodes.js
@@ -1,8 +1,11 @@
 const HTTP_STATUS = {
   SUCCESS: 200,
   CREATED: 201,
+  NO_CONTENT: 204,
   BAD_REQUEST: 400,
+  UNAUTHORIZED: 401,
   NOT_FOUND: 404,
+  CONFLICT: 409,
 };
 
 module.exports = HTTP_STATUS;
