+++ utils/validate/schemas/index.js
@@ -1,11 +1,11 @@
 const addContactValidator = require('./addContact');
 const updateContactValidator = require('./updateContact');
 const registrationValidator = require('./registration');
-const updateUser = require('./updateUser');
+const updateSubscriptionValidator = require('./updateSubscription');
 
 module.exports = {
   addContactValidator,
   updateContactValidator,
   registrationValidator,
-  updateUser,
+  updateSubscriptionValidator,
 };
